requirements.txt
absl-py==1.4.0
aiofiles==22.1.0
aiosqlite==0.18.0
ansitable==0.9.7
antlr4-python3-runtime==4.9.3
anyio==3.6.2
asttokens==2.2.1
async-generator==1.10
attrs==23.1.0
backcall==0.2.0
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
cmake==3.26.3
colorama==0.4.6
colored==2.2.3
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
easydict==1.10
exceptiongroup==1.1.1
executing==1.2.0
filelock==3.12.0
fonttools==4.39.4
freetype-py==2.4.0
google-auth==2.18.0
google-auth-oauthlib==0.4.6
grpcio==1.54.0
h11==0.14.0
HeapDict==1.0.1
idna==3.4
imageio==2.28.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
iopath==0.1.10
ipykernel==6.23.0
ipython==8.13.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json-tricks==3.17.3
jsonpatch==1.32
jsonpointer==2.3
jupyter_client==8.2.0
jupyter_core==5.3.0
kiwisolver==1.4.4
kornia==0.7.0
lit==16.0.3
lxml==4.9.2
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.6
networkx==2.5
numpy==1.22.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencv-python==4.7.0.72
outcome==1.2.0
packaging==23.1
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pgraph-python==0.6.2
pickleshare==0.7.5
Pillow==9.5.0
pinocchio==0.3
platformdirs==3.5.1
portalocker==2.8.2
progress==1.6
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybullet==3.2.5
pycocotools==2.0.7
pycollada==0.6
pyglet==2.0.7
Pygments==2.15.1
PyOpenGL==3.1.0
pyparsing==3.0.9
pyrender==0.1.45
python-dateutil==2.8.2
pytorch3d==0.7.4
pytz==2023.3
PyYAML==5.1
pyzmq==25.0.2
requests==2.30.0
requests-oauthlib==1.3.1
roboticstoolbox-python==1.0.1
rsa==4.9
rtb-data==1.0.1
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
shapely==2.0.1
simplejson==3.17.0
six==1.16.0
smplx==0.1.28
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
spatialgeometry==1.0.3
spatialmath-python==1.0.5
stack-data==0.6.2
swift-sim==1.0.1
sympy==1.12
tblib==1.7.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorboardX==2.6.2
termcolor==2.2.0
terminado==0.17.1
thop==0.1.1.post2209072238
threadpoolctl==3.1.0
tinycss2==1.2.1
tomli==2.0.1
toolz==0.12.0
torch==1.13.1+cu117
torch-summary==1.4.5
torch-utils==0.1.2
torchgeometry==0.1.2
torchnet==0.0.4
torchvision==0.14.1+cu117
tornado==6.3.1
tqdm==4.41.1
traitlets==5.9.0
transform3d==0.0.4
transforms3d==0.3.1
trimesh==3.18.1
trio==0.22.0
trio-websocket==0.9.2
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
urdfpy==0.0.22
uri-template==1.2.0
urllib3==1.26.14
visdom==0.2.3
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.0
websockets==11.0.3
Werkzeug==2.2.3
wget==3.2
wsproto==1.2.0
xarray==0.14.1
y-py==0.5.9
ypy-websocket==0.8.2
zict==2.2.0
zipp==3.15.0
